+++ RELEASES.md
@@ -9,6 +9,7 @@ Released on TBD (UTC).
 - Added Interactive Brokers cache config support for historical provider (#2942), thanks @ms32035
 - Added Interactive Brokers support for fetching orders from all clients (#2948), thanks @dinana
 - Added Bybit SPOT position reports with opt-in `use_spot_position_reports` config option for `BybitExecClientConfig`
+- Added OKX trade mode per order via `params` using `td_mode` key
 
 ### Breaking Changes
 None